feat: Normalize combat configuration schema and improve readability
- Refactored projectile configuration to use float for max lifetime.
- Updated bow aiming and firing configurations to use target angular speed instead of rotate speed.
- Introduced EnumBowShootType for clearer shoot type definitions in bow abilities.
- Created new XML definitions for equipment and marble abilities, ensuring clear separation of main and extension capabilities.
- Established consistent naming and type conventions for timing and motion fields across configurations.
- Enhanced factory assembly logic to ensure deterministic ability attachment and clear separation of main and extension capabilities.
- Added comprehensive documentation and specifications to guide future development and maintainability.

+++ UnityProject/Assets/GameScripts/HotFix/GameProto/GameConfig/Gameplay/Combat/ProjectileLifetimeConfig.cs
@@ -25,7 +25,7 @@ public sealed partial class ProjectileLifetimeConfig : Luban.BeanBase
         return new Gameplay.Combat.ProjectileLifetimeConfig(_buf);
     }
 
-    public readonly int? MaxLifetime;
+    public readonly float? MaxLifetime;
    
     public const int __ID__ = -2074794810;
     public override int GetTypeId() => __ID__;
